moved styled-components for result view into result directory
